Moved strictly related Twitter processing classes to twitter subpackage.
