+++ src/main/scala/ztis/twitter/KafkaTwitterStreamProducer.scala
@@ -1,7 +1,11 @@
-package ztis
+package ztis.twitter
 
 import com.typesafe.config.ConfigFactory
+import ztis.KafkaProducer
 
+/**
+ * Created by adebski on 19.04.15.
+ */
 object KafkaTwitterStreamProducer {
   val producer = new KafkaProducer(ConfigFactory.load("twitter-stream"))
 }
