+++ src/main/scala/ztis/twitter/Tweet.java
@@ -1,4 +1,4 @@
-package ztis;
+package ztis.twitter;
 
 import twitter4j.Status;
