+++ src/main/scala/ztis/twitter/TwitterLinkExtractorApp.scala
@@ -1,9 +1,13 @@
-package ztis
+package ztis.twitter
 
 import java.util.concurrent.Executors
 
 import com.typesafe.config.ConfigFactory
+import ztis.{CassandraClient, KafkaConsumer}
 
+/**
+ * Created by adebski on 19.04.15.
+ */
 object TwitterLinkExtractorApp extends App {
   val config = ConfigFactory.load("twitter-link-extractor")
   val executor = Executors.newFixedThreadPool(config.getInt("worker-threads"))
